move GEF UI constants into server directory
