+++ gef-ui/server/controllers/about-facility/index.js
@@ -2,9 +2,8 @@ const {
   add, format, isAfter, isBefore, set,
 } = require('date-fns');
 const api = require('../../services/api');
-const { FACILITY_TYPE } = require('../../../constants');
+const { FACILITY_TYPE, DATE_FORMAT } = require('../../constants');
 const { isTrueSet, validationErrorHandler } = require('../../utils/helpers');
-const CONSTANTS = require('../../../constants');
 
 const aboutFacility = async (req, res) => {
   const { params, query } = req;
@@ -247,8 +246,8 @@ const validateAboutFacility = async (req, res) => {
       name: body.facilityName,
       shouldCoverStartOnSubmission: isTrueSet(body.shouldCoverStartOnSubmission),
       monthsOfCover: body.monthsOfCover || null,
-      coverStartDate: coverStartDate ? format(coverStartDate, CONSTANTS.DATE_FORMAT.COVER) : null,
-      coverEndDate: coverEndDate ? format(coverEndDate, CONSTANTS.DATE_FORMAT.COVER) : null,
+      coverStartDate: coverStartDate ? format(coverStartDate, DATE_FORMAT.COVER) : null,
+      coverEndDate: coverEndDate ? format(coverEndDate, DATE_FORMAT.COVER) : null,
     });
 
     if (isTrueSet(saveAndReturn) || status === 'change') {
