+++ gef-ui/server/controllers/automatic-cover/index.js
@@ -1,5 +1,5 @@
 const { validationErrorHandler, stringToBoolean } = require('../../utils/helpers');
-const { DEAL_SUBMISSION_TYPE } = require('../../../constants');
+const { DEAL_SUBMISSION_TYPE } = require('../../constants');
 const { getValidationErrors, deriveCoverType } = require('./helpers');
 
 const api = require('../../services/api');
