+++ gef-ui/server/controllers/enter-exporters-correspondence-address/index.js
@@ -1,6 +1,6 @@
 const api = require('../../services/api');
 const { validationErrorHandler, isTrueSet } = require('../../utils/helpers');
-const { DEFAULT_COUNTRY } = require('../../../constants');
+const { DEFAULT_COUNTRY } = require('../../constants');
 
 const enterExportersCorrespondenceAddress = async (req, res) => {
   const { params, session, query } = req;
