+++ gef-ui/server/utils/user-authorisation-level.js
@@ -1,4 +1,4 @@
-const { PROGRESS, AUTHORISATION_LEVEL } = require('../../constants');
+const { PROGRESS, AUTHORISATION_LEVEL } = require('../constants');
 
 const statesWhereMakerHasEditAccess = [
   PROGRESS.DRAFT,
